+++ ci/ray_ci/BUILD.bazel
@@ -17,8 +17,8 @@ py_library(
     srcs = glob(
         ["*.py"],
         exclude = [
-            "test_*.py",
-            "ray_ci.py",
+            "tests/*.py",
+            "test_in_docker.py",
         ],
     ),
     visibility = ["//visibility:private"],
@@ -29,15 +29,17 @@ py_library(
 )
 
 py_binary(
-    name = "ray_ci",
-    srcs = ["ray_ci.py"],
-    deps = [":ray_ci_lib"],
+    name = "test_in_docker",
+    srcs = ["test_in_docker.py"],
+    deps = [
+        ":ray_ci_lib",
+    ],
 )
 
 py_test(
-    name = "test_runner",
+    name = "test_test_runner",
     size = "small",
-    srcs = ["test_runner.py"],
+    srcs = ["tests/test_test_runner.py"],
     tags = [
         "ci_unit",
         "team:ci",
@@ -52,7 +54,7 @@ py_test(
 py_test(
     name = "test_utils",
     size = "small",
-    srcs = ["test_utils.py"],
+    srcs = ["tests/test_utils.py"],
     tags = [
         "ci_unit",
         "team:ci",
@@ -64,9 +66,9 @@ py_test(
 )
 
 py_test(
-    name = "test_container",
+    name = "test_test_container",
     size = "small",
-    srcs = ["test_container.py"],
+    srcs = ["tests/test_test_container.py"],
     tags = [
         "ci_unit",
         "team:ci",
@@ -80,8 +82,8 @@ py_test(
 py_test(
     name = "test_bazel_sharding",
     size = "small",
-    srcs = ["test_bazel_sharding.py"],
-    data = ["mock_BUILD"],
+    srcs = ["tests/test_bazel_sharding.py"],
+    data = ["tests/mock_BUILD"],
     tags = [
         "ci_unit",
         "team:ci",
