+++ ci/ray_ci/test_container.py
@@ -1,51 +1,63 @@
+import os
+import subprocess
 import sys
-
-import pytest
-from unittest import mock
 from typing import List
 
-from ci.ray_ci.container import run_script_in_docker, run_tests
-from ci.ray_ci.utils import chunk_into_n
+from ci.ray_ci.utils import shard_tests
+from ci.ray_ci.container import Container
 
 
-class MockPopen:
+class TestContainer(Container):
     """
-    Mock subprocess.Popen. This process returns 0 if both test_targets and
-    commands are not empty; otherwise return 1.
+    A wrapper for running tests in ray ci docker container
     """
 
-    def __init__(self, test_targets: List[str]):
-        self.test_targets = test_targets
+    def __init__(self, team: str) -> None:
+        super().__init__(f"{team}build")
 
-    def wait(self) -> int:
-        return 0 if self.test_targets else 1
+    def run_tests(self, test_targets: List[str], parallelism) -> bool:
+        """
+        Run tests parallelly in docker.  Return whether all tests pass.
+        """
+        chunks = [shard_tests(test_targets, parallelism, i) for i in range(parallelism)]
+        runs = [self._run_tests_in_docker(chunk) for chunk in chunks]
+        exits = [run.wait() for run in runs]
+        return all(exit == 0 for exit in exits)
 
+    def setup_test_environment(self) -> None:
+        """
+        Build the docker image for running tests
+        """
+        env = os.environ.copy()
+        env["DOCKER_BUILDKIT"] = "1"
+        subprocess.check_call(
+            [
+                "docker",
+                "build",
+                "--build-arg",
+                f"BASE_IMAGE={self._get_docker_image()}",
+                "-t",
+                self._get_docker_image(),
+                "-f",
+                "/ray/ci/ray_ci/tests.env.Dockerfile",
+                "/ray",
+            ],
+            env=env,
+            stdout=sys.stdout,
+            stderr=sys.stderr,
+        )
 
-def test_run_script_in_docker() -> None:
-    def _mock_check_output(input: List[str]) -> None:
-        input_str = " ".join(input)
-        assert "/bin/bash -ice run command" in input_str
-
-    with mock.patch("subprocess.check_output", side_effect=_mock_check_output):
-        run_script_in_docker("run command", "team")
-
-
-def test_run_tests() -> None:
-    def _mock_run_tests_in_docker(test_targets: List[str], team: str) -> MockPopen:
-        return MockPopen(test_targets)
-
-    def _mock_shard_tests(tests: List[str], workers: int, worker_id: int) -> List[str]:
-        return chunk_into_n(tests, workers)[worker_id]
-
-    with mock.patch(
-        "ci.ray_ci.container._run_tests_in_docker",
-        side_effect=_mock_run_tests_in_docker,
-    ), mock.patch("ci.ray_ci.container.shard_tests", side_effect=_mock_shard_tests):
-        # test_targets are not empty
-        assert run_tests("team", ["t1", "t2"], 2)
-        # test_targets is empty after chunking
-        assert not run_tests("team", ["t1"], 2)
-
-
-if __name__ == "__main__":
-    sys.exit(pytest.main(["-v", __file__]))
+    def _run_tests_in_docker(self, test_targets: List[str]) -> subprocess.Popen:
+        commands = []
+        if os.environ.get("BUILDKITE_BRANCH", "") == "master":
+            commands.extend(
+                [
+                    "cleanup() { ./ci/build/upload_build_info.sh; }",
+                    "trap cleanup EXIT",
+                ]
+            )
+        commands.append(
+            "bazel test --config=ci $(./ci/run/bazel_export_options) "
+            f"{' '.join(test_targets)}",
+        )
+        return subprocess.Popen(self._get_run_command("\n".join(commands)))
